pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "huapir"
version = "3.3.0a2"
authors = [
    { name="Lss233", email="i@lss233.com" },
]
description = "A framework for building AI agents"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "pydantic>=2.0.0",
    "pydantic-core>=2.27.2",
    "ruamel.yaml",
    "pytest",
    "pytest-asyncio",
    "python-telegram-bot",
    "telegramify-markdown",
    "loguru",
    "aiohttp",
    "requests",
    "quart>=0.18.4",
    "quart-cors>=0.7.0",
    "fastapi>=0.110.0",
    "wechatpy",
    "pycryptodome",
    "redis[hiredis]",
    "bcrypt>=4.0.1",
    "PyJWT>=2.8.0",
    "hypercorn>=0.15.0",
    "psutil>=5.9.0",
    "setuptools",
    "tomli>=2.0.0",
    "pre-commit",
    "curl_cffi",
    "python-magic ; platform_system != 'Windows'",
    "python-magic-bin ; platform_system == 'Windows'",
    "ymbotpy",
    "Pillow",
    "pytz",
    "sqlalchemy",
    "alembic",
    "pygls",
    "jedi",
    "pyflakes",
]


[project.scripts]
huapir = "huapir.__main__:main"

[project.urls]
"Homepage" = "https://github.com/lss233/chatgpt-mirai-qq-bot"
"Bug Tracker" = "https://github.com/lss233/chatgpt-mirai-qq-bot/issues"
"Documentation" = "https://kirara-docs.app.lss233.com"

[tool.setuptools.packages.find]
where = ["."]
include = ["huapir*"]
namespaces = true

[[tool.uv.index]]
url = "https://mirrors.ustc.edu.cn/pypi/simple"
default = true

[options]
include_package_data = true

[tool.mypy]
disable_error_code = "override, assignment, annotation-unchecked, import-untyped"
